zoomerjoin: Superlatively Fast Fuzzy Joins

Empowers users to fuzzily-merge data frames with millions or tens of millions of rows in minutes with low memory usage. The package uses the locality sensitive hashing algorithms developed by Datar, Immorlica, Indyk and Mirrokni (2004) doi:10.1145/997817.997857 , and Broder (1998) doi:10.1109/SEQUEN.1997.666900 to avoid having to compare every pair of records in each dataset, resulting in fuzzy-merges that finish in linear time.

See also

Author

Maintainer: Beniamino Green beniamino.green@yale.edu [copyright holder]

Other contributors:

Etienne Bacher etienne.bacher@protonmail.com (ORCID) [contributor]

The authors of the dependency Rust crates (see inst/AUTHORS file for details) [contributor, copyright holder]